Refactor coral catalog APIs to SPI model
Introduce a 3-module catalog SPI architecture that decouples Hive and
Iceberg implementations from the core classpath:

- coral-catalog-spi: Pure interfaces (CoralCatalog, CoralTable, TableType),
  type system classes, and new SPI contracts (CoralCalciteTableAdapterFactory,
  CoralCalciteTableAdapterRegistry) with zero Hive/Iceberg dependencies.

- coral-catalog-hive: Hive catalog implementations (HiveTable,
  HiveCalciteTableAdapter, HiveCalciteViewAdapter, TypeConverter) with
  SPI factory registration via META-INF/services.

- coral-catalog-iceberg: Iceberg catalog implementations (IcebergTable,
  IcebergCalciteTableAdapter, IcebergHiveTableConverter) with SPI factory
  registration via META-INF/services.

coral-common is refactored to use CoralCalciteTableAdapterRegistry for
SPI-based dispatch instead of instanceof checks. Original implementation
classes are retained as @Deprecated wrappers for backward compatibility.
Hive/Hadoop/Iceberg dependencies are demoted from api to implementation
scope.

Known limitation: ToRelConverter.toSqlNode() still couples the core to
Hive's Table type (issue #575).
