+++ coral-catalog-hive/src/main/resources/META-INF/services/com.linkedin.coral.common.catalog.CoralCalciteTableAdapterFactory
@@ -1,0 +1,1 @@
+com.linkedin.coral.catalog.hive.HiveCalciteTableAdapterFactory
