+++ coral-catalog-spi/src/main/java/com/linkedin/coral/common/catalog/CoralTable.java
