+++ coral-catalog-spi/src/main/java/com/linkedin/coral/common/catalog/TableType.java
