+++ coral-catalog-spi/src/main/java/com/linkedin/coral/common/types/CharType.java
