+++ coral-common/src/main/java/com/linkedin/coral/common/HiveCalciteViewAdapter.java
@@ -7,70 +7,31 @@ package com.linkedin.coral.common;
 
 import java.util.List;
 
-import org.apache.calcite.plan.RelOptTable;
-import org.apache.calcite.rel.RelNode;
-import org.apache.calcite.rel.RelRoot;
-import org.apache.calcite.schema.TranslatableTable;
 import org.apache.hadoop.hive.metastore.api.Table;
 
-import com.linkedin.coral.com.google.common.base.Throwables;
-import com.linkedin.coral.com.google.common.collect.ImmutableList;
 import com.linkedin.coral.common.catalog.HiveTable;
 
 
 /**
- * Calcite adapter for Hive views, extending HiveCalciteTableAdapter with TranslatableTable support
- * for recursive expansion of view definitions.
- *
- * <p>This adapter enables Calcite to process Hive views by implementing the TranslatableTable interface,
- * which allows the view definition to be expanded and converted into a relational algebra tree.
- *
- * <p><b>Integration with ParseTreeBuilder:</b> This class inherits the Dali UDF metadata extraction
- * methods from {@link HiveCalciteTableAdapter}, which are used by {@code ParseTreeBuilder} and
- * {@code HiveFunctionResolver} for parsing view definitions with custom functions. These components
- * are currently tightly coupled to {@link org.apache.hadoop.hive.metastore.api.Table} and cannot
- * work directly with {@link com.linkedin.coral.common.catalog.CoralTable}.
- *
- * <p>This coupling is being addressed in <a href="https://github.com/linkedin/coral/issues/575">issue #575</a>,
- * which will refactor these APIs to accept {@code CoralTable} instead, enabling multi-format support.
- *
- * @see HiveCalciteTableAdapter
- * @see <a href="https://github.com/linkedin/coral/issues/575">Issue #575: Refactor ParseTreeBuilder to Use CoralTable</a>
+ * @deprecated Use {@link com.linkedin.coral.catalog.hive.HiveCalciteViewAdapter} instead.
+ * This class is retained for backward compatibility and will be removed in a future release.
  */
-public class HiveCalciteViewAdapter extends HiveCalciteTableAdapter implements TranslatableTable {
-  private final List<String> schemaPath;
+@Deprecated
+public class HiveCalciteViewAdapter extends com.linkedin.coral.catalog.hive.HiveCalciteViewAdapter {
 
   /**
-   * Constructor to create bridge from hive table to calcite table
-   *
-   * @param hiveTable Hive table
-   * @param schemaPath Calcite schema path
+   * @deprecated Use {@link com.linkedin.coral.catalog.hive.HiveCalciteViewAdapter#HiveCalciteViewAdapter(Table, List)} instead.
    */
+  @Deprecated
   public HiveCalciteViewAdapter(Table hiveTable, List<String> schemaPath) {
-    super(hiveTable);
-    this.schemaPath = schemaPath;
+    super(hiveTable, schemaPath);
   }
 
   /**
-   * Constructor accepting HiveCoralTable for unified catalog integration.
-   *
-   * @param coralTable HiveCoralTable from catalog
-   * @param schemaPath Calcite schema path
+   * @deprecated Use {@link com.linkedin.coral.catalog.hive.HiveCalciteViewAdapter#HiveCalciteViewAdapter(com.linkedin.coral.catalog.hive.HiveTable, List)} instead.
    */
+  @Deprecated
   public HiveCalciteViewAdapter(HiveTable coralTable, List<String> schemaPath) {
-    super(coralTable);
-    this.schemaPath = schemaPath;
-  }
-
-  @Override
-  public RelNode toRel(RelOptTable.ToRelContext relContext, RelOptTable relOptTable) {
-    try {
-      RelRoot root = relContext.expandView(relOptTable.getRowType(), hiveTable.getViewExpandedText(), schemaPath,
-          ImmutableList.of(hiveTable.getTableName()));
-      return root.rel;
-    } catch (Exception e) {
-      Throwables.propagateIfInstanceOf(e, RuntimeException.class);
-      throw new RuntimeException("Error while parsing view definition", e);
-    }
+    super(coralTable.getHiveTable(), schemaPath);
   }
 }
